+++ order-service/app/api/routes.py
@@ -659,11 +659,12 @@ from models.order_model import (
     TransactionTypeEnum,
     MovementTypeEnum,
 )
-from utils.auth import verify_token, record_transaction, record_partial_transaction
+from utils.auth import verify_token
+from utils.transaction import record_transaction, record_partial_transaction
 from models.saas_context import SaasContext
 from typing import Optional
 from entity.inventory_entity import InventoryEntity
-from services.order_service import (
+from ..services.order_service import (
     _root_dinein_id,
     _order_row_to_flat,
     STATUS_PRIORITY,
